+++ spring-core/src/test/java/org/springframework/util/NumberUtilsTests.java
@@ -372,15 +372,15 @@ public class NumberUtilsTests {
 	}
 
 	private void assertIntegerEquals(String anInteger) {
-		assertThat(NumberUtils.parseNumber(anInteger, Integer.class).intValue()).as("Integer did not parse").isEqualTo((long) Integer.MAX_VALUE);
+		assertThat(NumberUtils.parseNumber(anInteger, Integer.class).intValue()).as("Integer did not parse").isEqualTo(Integer.MAX_VALUE);
 	}
 
 	private void assertShortEquals(String aShort) {
-		assertThat(NumberUtils.parseNumber(aShort, Short.class).shortValue()).as("Short did not parse").isEqualTo((long) Short.MAX_VALUE);
+		assertThat(NumberUtils.parseNumber(aShort, Short.class).shortValue()).as("Short did not parse").isEqualTo(Short.MAX_VALUE);
 	}
 
 	private void assertByteEquals(String aByte) {
-		assertThat(NumberUtils.parseNumber(aByte, Byte.class).byteValue()).as("Byte did not parse").isEqualTo((long) Byte.MAX_VALUE);
+		assertThat(NumberUtils.parseNumber(aByte, Byte.class).byteValue()).as("Byte did not parse").isEqualTo(Byte.MAX_VALUE);
 	}
 
 	private void assertNegativeLongEquals(String aLong) {
@@ -388,15 +388,15 @@ public class NumberUtilsTests {
 	}
 
 	private void assertNegativeIntegerEquals(String anInteger) {
-		assertThat(NumberUtils.parseNumber(anInteger, Integer.class).intValue()).as("Integer did not parse").isEqualTo((long) Integer.MIN_VALUE);
+		assertThat(NumberUtils.parseNumber(anInteger, Integer.class).intValue()).as("Integer did not parse").isEqualTo(Integer.MIN_VALUE);
 	}
 
 	private void assertNegativeShortEquals(String aShort) {
-		assertThat(NumberUtils.parseNumber(aShort, Short.class).shortValue()).as("Short did not parse").isEqualTo((long) Short.MIN_VALUE);
+		assertThat(NumberUtils.parseNumber(aShort, Short.class).shortValue()).as("Short did not parse").isEqualTo(Short.MIN_VALUE);
 	}
 
 	private void assertNegativeByteEquals(String aByte) {
-		assertThat(NumberUtils.parseNumber(aByte, Byte.class).byteValue()).as("Byte did not parse").isEqualTo((long) Byte.MIN_VALUE);
+		assertThat(NumberUtils.parseNumber(aByte, Byte.class).byteValue()).as("Byte did not parse").isEqualTo(Byte.MIN_VALUE);
 	}
 
 	private void assertToNumberOverflow(Number number, Class<? extends Number> targetClass) {
