+++ spring-web/src/test/java/org/springframework/http/converter/json/MappingJackson2HttpMessageConverterTests.java
@@ -91,7 +91,7 @@ public class MappingJackson2HttpMessageConverterTests {
 		MyBean result = (MyBean) converter.read(MyBean.class, inputMessage);
 		assertThat(result.getString()).isEqualTo("Foo");
 		assertThat(result.getNumber()).isEqualTo(42);
-		assertThat(result.getFraction()).isCloseTo((double) 42F, within((double) 0F));
+		assertThat(result.getFraction()).isCloseTo(42F, within(0F));
 
 		assertThat(result.getArray()).isEqualTo(new String[] {"Foo", "Bar"});
 		assertThat(result.isBool()).isTrue();
@@ -222,7 +222,7 @@ public class MappingJackson2HttpMessageConverterTests {
 		MyBean result = results.get(0);
 		assertThat(result.getString()).isEqualTo("Foo");
 		assertThat(result.getNumber()).isEqualTo(42);
-		assertThat(result.getFraction()).isCloseTo((double) 42F, within((double) 0F));
+		assertThat(result.getFraction()).isCloseTo(42F, within(0F));
 
 		assertThat(result.getArray()).isEqualTo(new String[] {"Foo", "Bar"});
 		assertThat(result.isBool()).isTrue();
@@ -254,7 +254,7 @@ public class MappingJackson2HttpMessageConverterTests {
 		MyBean result = results.get(0);
 		assertThat(result.getString()).isEqualTo("Foo");
 		assertThat(result.getNumber()).isEqualTo(42);
-		assertThat(result.getFraction()).isCloseTo((double) 42F, within((double) 0F));
+		assertThat(result.getFraction()).isCloseTo(42F, within(0F));
 
 		assertThat(result.getArray()).isEqualTo(new String[] {"Foo", "Bar"});
 		assertThat(result.isBool()).isTrue();
@@ -287,7 +287,7 @@ public class MappingJackson2HttpMessageConverterTests {
 		MyBean result = results.get(0);
 		assertThat(result.getString()).isEqualTo("Foo");
 		assertThat(result.getNumber()).isEqualTo(42);
-		assertThat(result.getFraction()).isCloseTo((double) 42F, within((double) 0F));
+		assertThat(result.getFraction()).isCloseTo(42F, within(0F));
 
 		assertThat(result.getArray()).isEqualTo(new String[] {"Foo", "Bar"});
 		assertThat(result.isBool()).isTrue();
